+++ src/components/form/model/renderFormBlock.tsx
@@ -1,6 +1,6 @@
 import {ReactElement} from 'react';
 import {Control, FieldValues} from 'react-hook-form';
-import {InputWithController, PasswordInput} from '@/shared/ui';
+import {InputWithController} from '@/shared/ui';
 import {FormOption, FormVariantsEnum} from './types';
 
 type Props<T extends FieldValues> = {
@@ -15,14 +15,5 @@ export const renderFormBlock = <T extends FieldValues>({
   switch (option.variant) {
     case FormVariantsEnum.Input:
       return <InputWithController key={option.id} option={option} control={control} />;
-    case FormVariantsEnum.PasswordInput:
-      return (
-        <InputWithController
-          key={option.id}
-          InputComponent={PasswordInput}
-          option={option}
-          control={control}
-        />
-      );
   }
 };
