+++ src/components/header/ui/Header.tsx
@@ -1,11 +1,24 @@
-import {AppBar} from '@mui/material';
+import {AppBar, Stack} from '@mui/material';
 import {ReactElement} from 'react';
-import {Navbar} from './Navbar';
+import {HeaderNavigation} from './HeaderNavigation';
+import {HeaderToolbar} from './HeaderToolbar';
 
 export const Header = (): ReactElement => {
   return (
     <AppBar position="static">
-      <Navbar />
+      <Stack
+        sx={(theme) => ({
+          [theme.breakpoints.down('md')]: {
+            flexDirection: 'row',
+            alignItems: 'center'
+          },
+          flexDirection: 'column',
+          alignItems: 'none'
+        })}
+      >
+        <HeaderNavigation />
+        <HeaderToolbar />
+      </Stack>
     </AppBar>
   );
 };
