+++ src/shared/const/errorMessages.const.ts
@@ -2,6 +2,8 @@ export const ErrorMessages = {
   ERROR: 'Error',
   EMAIL_REQUIRED: 'Email is required.',
   INVALID_EMAIL: 'Email must be a valid email.',
+  PASSWORD_ERROR_MESSAGE:
+    'Password must contain at least 2 lowercase, 2 uppercase, 2 digits, and 2 special characters',
   PASSWORD_REQUIRED: 'Password is required.',
   IS_REQUIRED: (label: string) => `${label} is required.`,
   AT_LEAST_LENGTH: (length: number, label: string) =>
@@ -10,5 +12,5 @@ export const ErrorMessages = {
     `${label} must be at most ${length} characters.`,
   MIN_LENGTH: ({length, label}: {length: number; label?: string}) =>
     `${label ? label : 'Minimum'} length should be ${length} characters.`,
-  INVALID_PHONE_FORMAT: (length?: number) => `Phone must contain ${length ? length : 12} digits.`
+  INVALID_PHONE_FORMAT: (length: number = 12) => `Phone must contain ${length} digits.`
 };
