+++ src/shared/const/regex.conts.ts
@@ -1,0 +1,2 @@
+export const passwordRegex =
+  /^(?=.*[a-z].*[a-z])(?=.*[A-Z].*[A-Z])(?=.*\d.*\d)(?=.*[^a-zA-Z0-9].*[^a-zA-Z0-9]).*$/;
