+++ src/shared/ui/index.ts
@@ -1,4 +1,4 @@
-export {AuthButton} from './button/AuthButton.tsx';
+export {AuthMenuButton} from './button/AuthMenuButton.tsx';
 export {Button} from './button/Button.tsx';
 export {CartButton} from './button/CartButton.tsx';
 export {MenuButton} from './button/MenuButton.tsx';
@@ -6,6 +6,5 @@ export {ThemeButton} from './button/ThemeButton';
 export {Container} from './container/Container.tsx';
 export {Input} from './input/Input.tsx';
 export {InputWithController} from './input/InputWithController.tsx';
-export {PasswordInput} from './input/PasswordInput.tsx';
 export {AuthModal} from './modal/AuthModal.tsx';
 export {Modal} from './modal/Modal.jsx';
