+++ src/shared/ui/input/InputWithController.tsx
@@ -2,25 +2,22 @@ import {ReactElement} from 'react';
 import {Control, Controller, FieldValues, Path} from 'react-hook-form';
 import {FormOption, FormVariantsEnum} from '@/components/form';
 import {Input} from './Input';
-import {PasswordInput} from './PasswordInput';
 
 interface Props<T extends FieldValues> {
   option: FormOption<FormVariantsEnum>;
   control: Control<T>;
-  InputComponent?: typeof Input | typeof PasswordInput;
 }
 
 export const InputWithController = <T extends FieldValues>({
   option,
-  control,
-  InputComponent = Input
+  control
 }: Props<T>): ReactElement => {
   return (
     <Controller
       control={control}
       name={option.id as Path<T>}
       render={({field, fieldState: {error}}) => (
-        <InputComponent
+        <Input
           key={option.id}
           type={option.type}
           label={option.name}
